refactor(router): consolidate error handling by removing unused error classes
- Deleted several error classes related to routing, including ControllerMethodNotFoundError, ControllerRegistrationError, DomainMismatchError, and others, to streamline error management.
- Updated the index file to reflect the removal of these classes, enhancing maintainability and clarity in the error handling structure.
