+++ packages/core/src/router/__tests__/response-validation.spec.ts
@@ -1,17 +1,17 @@
-import { createMock } from '@stratal/testing/mocks'
 import { OpenAPIHono } from '@hono/zod-openapi'
+import { createMock } from '@stratal/testing/mocks'
 import { describe, expect, it } from 'vitest'
-import { z } from '../../i18n/validation'
+import { ERROR_CODES } from '../../errors/error-codes'
 import type { ZodType } from '../../i18n/validation'
+import { z } from '../../i18n/validation'
 import type { LoggerService } from '../../logger/services/logger.service'
 import type { ModuleRegistry } from '../../module/module-registry'
-import { ERROR_CODES } from '../../errors/error-codes'
-import { ResponseValidationError } from '../errors/response-validation.error'
+import { ResponseValidationError } from '../errors'
 import type { HonoApp } from '../hono-app'
 import { RouteRegistry } from '../route-registry'
+import type { LocalePathService } from '../services/locale-path.service'
 import { RouteRegistrationService } from '../services/route-registration.service'
 import type { VersioningService } from '../services/versioning.service'
-import type { LocalePathService } from '../services/locale-path.service'
 import type { RouteConfig, RouterEnv } from '../types'
 
 const mockLogger = createMock<LoggerService>()
