+++ packages/core/src/router/__tests__/router.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from 'vitest'
 import { z } from '../../i18n/validation'
 import type { Constructor } from '../../types'
-import { RouterUseScopeError } from '../errors/router-use-scope.error'
+import { RouterUseScopeError } from '../errors'
 import type { Middleware } from '../middleware.interface'
 import { Router } from '../router'
 import * as internal from '../router.internals'
@@ -13,10 +13,10 @@ class AdminMiddleware { handle() { /**/ } }
 class AuditMiddleware { handle() { /**/ } }
 
 // Stub controllers
-class UsersController {}
-class PostsController {}
-class AdminController {}
-class HealthController {}
+class UsersController { }
+class PostsController { }
+class AdminController { }
+class HealthController { }
 
 describe('Router', () => {
   describe('fluent API', () => {
