+++ packages/core/src/router/hono-app.ts
@@ -11,9 +11,7 @@ import { LOGGER_TOKENS, type LoggerService } from '../logger'
 import { OPENAPI_TOKENS, type OpenAPIService } from '../openapi'
 import type { Constructor } from '../types'
 import { ROUTER_CONTEXT_KEYS } from './constants'
-import { RouteNotFoundError } from './errors'
-import { HonoAppAlreadyConfiguredError } from './errors/hono-app-already-configured.error'
-import { SchemaValidationError } from './errors/schema-validation.error'
+import { HonoAppAlreadyConfiguredError, RouteNotFoundError, SchemaValidationError } from './errors'
 import { createLoggerMiddleware, createMiddlewareChain } from './middleware'
 import type { Middleware } from './middleware.interface'
 import { RouterContext } from './router-context'
